[Doc] Improved documentation across the Display module
* Rewrote class descriptions for Bitmap, Clipboard, Display, Pointer and Surface with clearer language and more accurate detail
* Updated field and method documentation with precise error codes, corrected descriptions and cross-references
* Fixed ResizeFeedback and DisplayType field comments which had incorrect placeholder content
* [AI] Updated doc-enhancer skill to verify error code completeness and avoid redundant error descriptions

+++ src/display/defs.h
@@ -373,7 +373,7 @@ class extDisplay : public objDisplay {
    #else
       APTR   WindowHandle;
    #endif
-   int16_t  Opacity;
+   float Opacity;
    char  Manufacturer[60];
    char  Chipset[40];
    char  Display[32];
